+++ src/constants.js
@@ -1,4 +1,4 @@
-export const VERSION = '2.0.10';
+export const VERSION = '2.1.0';
 
 export const SSH_COLORS = [
   'cyan',